rwengine: iwyu: reduce warnings in render subdirectory
