+++ rwengine/src/engine/GameData.cpp
@@ -5,6 +5,7 @@
 #include <cmath>
 #include <cstdlib>
 #include <fstream>
+#include <iostream>
 #include <iterator>
 #include <stdexcept>
