+++ rwengine/src/render/WaterRenderer.hpp
@@ -1,7 +1,15 @@
-#pragma once
+#ifndef _RWENGINE_WATERRENDERER_HPP_
+#define _RWENGINE_WATERRENDERER_HPP_
+
+#include <cstdint>
+#include <memory>
+#include <vector>
+
+#include <gl/DrawBuffer.hpp>
+#include <gl/GeometryBuffer.hpp>
+#include <gl/gl_core_3_3.h>
 
 #include <render/OpenGLRenderer.hpp>
-#include <rw/types.hpp>
 
 class GameRenderer;
 class GameWorld;
@@ -47,3 +55,5 @@ private:
     GLuint fbOutput;
     GLuint dataTexture;
 };
+
+#endif
